Code refactor changes
